requirements.txt
torch
numpy
Levenshtein
Pillow
torchvision
transformers
tensorboard